feat: Head, ScrollToTop Settings 폴더로 분리
